# docs/api-reference/swagger.yaml
openapi: "3.0.3"
info:
  version: 1.0.2
  title: Scheduler
  license:
    name: MIT
servers:
  - url: http://localhost:8080/api/v1
paths:
...
  /schedule:
    post:
      summary: Request a deployment to be scheduled on one of all available clouds matching selector (predicates with labels) and preferences (priorities with labels).
      operationId: schedule
      tags:
        - schedule
      requestBody:
        description: JSON object to specify selectors, priorities and tolerations
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleQuery"
      responses:
        '200':
          description: The Placement given by the scheduler, containing the target cloud that can be used.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Placement"
        '400':
          description: The service identified by uuid already has a placement.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '404':
          description: NotFound
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
    Cloud:
      description: Object that defines a Cloud.
      type: object
      required:
        - name
        - labels
      properties:
        name:
          type: string
        labels:
          type: object
          additionalProperties:
            type: string
        taints:
          type: array
          items:
            $ref: "#/components/schemas/Taint"
...
    UUID:
      type: string
      description: The UUID of the service in CloudForms. It used to identify placement and can be used later to retrieve a placement.
      format: uuid
    Annotations:
      description: Key / Value map to provide optional information. Annotations can be used to filter or identify objects.
      type: object
      additionalProperties:
        type: string
        minLength: 1
    Placement:
      type: object
      description: A placement is a record of what(uuid) / where(cloud) / when(date) something was scheduled.
      required:
        - uuid
        - cloud
        - creation_timestamp
        - annotations
      properties:
        uuid:
          $ref: "#/components/schemas/UUID"
        cloud:
          $ref: "#/components/schemas/Cloud"
        creation_timestamp:
          description: The date (UTC and RFC3339 format) the placement was made.
          type: string
          format: date-time
        annotations:
          $ref: "#/components/schemas/Annotations"
...
    Taint:
      type: object
      description: The cloud this taint is attached to has the "effect" on any deployment that does not tolerate the Taint.
      required:
        - key
        - effect
      properties:
        key:
          type: string
          description: The name of the taint to be applied to a cloud.
          example: memory-pressure
        value:
          type: string
          description: The value of the taint.
          example: critical
        effect:
          type: string
          description: |-
            The effect of the taint on deployments that do not tolerate the taint. <br /><br />
            NoSchedule: affects the predicates, Cloud can't be selected unless there is a matching Toleration.
            PreferNoSchedule: affects the priorities, Cloud will be de-prioritized.
          enum:
            - NoSchedule
            - PreferNoSchedule
    Toleration:
      type: object
      description: Object passed when requesting schedule to ignore taints.
      properties:
        key:
          type: string
          description: |-
            Key is the taint key that the toleration applies to.
            Empty means match all taint keys.
            If the key is empty, operator must be Exists;
            this combination means to match all values and all keys.
            (optional)
          example: memory-pressure
        operator:
          description: |-
            Operator represents a key's relationship to the value.
            Valid operators are Exists and Equal. Defaults to Equal.
            Exists is equivalent to wildcard for value, so that a cloud can
            tolerate all taints of a particular category.
            (optional)<br /><br />
            Equal: The key/value/effect parameters must match. This is the default.<br />
            Exists: The key/value/effect parameters must match. This is the default.
          enum:
            - Exists
            - Equal
          default: Equal
          type: string
        value:
          type: string
          description: |-
            Value is the taint value the toleration matches to.
            If the operator is Exists, the value should be empty, otherwise just a regular string.
            +optional
          example: critical
        effect:
          type: string
          description: |-
            Effect indicates the taint effect to match. Empty means match all taint effects.
            When specified, allowed values are NoSchedule, PreferNoSchedule.
            +optional
          enum:
            - NoSchedule
            - PreferNoSchedule
    ScheduleQuery:
      type: object
      required:
        - uuid
      properties:
        uuid:
          $ref: "#/components/schemas/UUID"
        cloud_selector:
          description: This dictionary describes the labels (key:value) that must be present in the clouds in order to be selected by the scheduler.
          type: object
          additionalProperties:
            type: string
        cloud_preference:
          type: object
          description: This dictionary describes the labels (key:value) that you would like to be present in the clouds in order to be selected by the scheduler. They change the priority and thus the clouds matching those labels will be selected first.
          additionalProperties:
            type: string
        tolerations:
          type: array
          description: The list of tolerations for this request. Any taint matching a toleration will be ignored (all taints ignored == cloud can be selected).
          items:
            $ref: "#/components/schemas/Toleration"
        annotations:
          $ref: "#/components/schemas/Annotations"
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string